moved functions from server.py to virtualmachine.py
also fixed a bug where you couldn't stop a vm in vmware player
